@@ -14,8 +14,8 @@ from schema import Optional, Or
 
 from organize.utils import SimulationFS, Template, safe_description
 
-from .action import Action
 from ._conflict_resolution import CONFLICT_OPTIONS, check_conflict, dst_from_options
+from .action import Action
 
 
 # this is taken from my PR
@@ -140,7 +140,7 @@ class Move(Action):
     ) -> None:
         if on_conflict not in CONFLICT_OPTIONS:
             raise ValueError(
-                "on_conflict must be one of %s" % ", ".join(CONFLICT_OPTIONS)
+                f"on_conflict must be one of {', '.join(CONFLICT_OPTIONS)}"
             )
 
         self.dest = Template.from_string(dest)
@@ -186,7 +186,7 @@ class Move(Action):
                 dst_fs = SimulationFS(dst_fs)
 
         if not skip:
-            self.print("Move to %s" % safe_description(dst_fs, dst_path))
+            self.print(f"Move to {safe_description(dst_fs, dst_path)}")
             if not simulate:
                 dst_fs.makedirs(dirname(dst_path), recreate=True)
                 move_action(src_fs, src_path, dst_fs, dst_path)
@@ -198,4 +198,4 @@ class Move(Action):
         }
 
     def __str__(self) -> str:
-        return "Move(dest=%s, conflict_mode=%s)" % (self.dest, self.conflict_mode)
+        return f"Move(dest={self.dest}, conflict_mode={self.conflict_mode})"
