@@ -122,7 +122,7 @@ def test_copy_into_dir_subfolders(testfiles):
     ],
 )
 def test_copy_conflict(testfs: FS, mode, files, test_txt_content):
-    config = """
+    config = f"""
     rules:
       - locations: "/"
         filters:
@@ -130,10 +130,8 @@ def test_copy_conflict(testfs: FS, mode, files, test_txt_content):
         actions:
           - copy:
               dest: "test.txt"
-              on_conflict: {}
-    """.format(
-        mode
-    )
+              on_conflict: {mode}
+    """
     testfs.writetext("file.txt", "new")
     testfs.writetext("test.txt", "old")
     core.run(config, simulate=False, working_dir=testfs)
